+++ R/api-reports.R
@@ -19,11 +19,10 @@ report <- function(connection, report_id, verbose=FALSE){
   return(response)
 }
 
-
 # Get the results of a newly created report instance.
 # This in-memory report instance can be used by other requests.
 # Returns complete HTTP response object.
-report_instance <- function(connection, report_id, offset=0, limit=1000, body = NULL, verbose=FALSE){
+report_instance <- function(connection, report_id, body=NULL, offset=0, limit=1000, verbose=FALSE){
   response <- httr::POST(url = paste0(connection@base_url, '/reports/', report_id, '/instances'),
                          add_headers('X-MSTR-AuthToken' = connection@auth_token,
                                      'X-MSTR-ProjectID' = connection@project_id),
@@ -37,7 +36,6 @@ report_instance <- function(connection, report_id, offset=0, limit=1000, body = 
   }
   return(response)
 }
-
 
 # Get the results of a previously created report instance, using the in-memory report instance
 # created by a POST /reports/{reportId}/instances request.
@@ -58,10 +56,12 @@ report_instance_id <- function(connection, report_id, instance_id, offset=0, lim
 
 # Get elements of a specific attribute of a specific report.
 # Returns complette HTTP response object.
-report_single_attr_elements <- function(connection, report_id, attribute_id, verbose = FALSE) {
+report_elements <- function(connection, report_id, attribute_id, offset=0, limit=25000, verbose=FALSE) {
   response <- httr::GET(url = paste0(connection@base_url, '/reports/', report_id, '/attributes/', attribute_id, '/elements'),
                         add_headers('X-MSTR-AuthToken' = connection@auth_token,
                                     'X-MSTR-ProjectID' = connection@project_id),
+                        query=list(offset = format(offset, scientific=FALSE, trim=TRUE),
+                                   limit = format(limit, scientific=FALSE, trim=TRUE)),
                         set_cookies(connection@cookies))
   if(verbose){
     print(response$url)