+++ R/cube.R
@@ -1,6 +1,7 @@
 # cube.R
 # Create and interact with MicroStrategy cubes
 
+
 #' @title Extract a MicroStrategy cube into a R Data.Frame
 #'
 #' @description Access, filter, publish, and extract data from MicroStrategy in-memory cubes
@@ -8,22 +9,34 @@
 #' @field cube_id Identifier of a cube.
 #' @examples
 #' \donttest{
+#' # Create a connection object.
+#' connection = connect_mstr(base_url, username, password, project_name)
+#' 
+#' # Create a cube object.
 #' my_cube <- Cube$new(connection=conn, cube_id="...")
-#' df <- my_cube$to_dataframe()
-#' # Use object IDs for metrics, attributes, and attribute elements to filter contents of a cube.
-#' my_cube$metrics
+#' 
+#' # See attributes and metrics in the report.
 #' my_cube$attributes
+#' my_cube$metrics
 #' my_cube$attr_elements
 #' 
-#' # Then, choose those elements by passing their IDs to the Cube.apply_filters() method. To see 
-#' # the chosen elements, call my_cube.filters and to clear any active filters, 
-#' # call my_cube.clear_filters().
-#' my_cube$apply_filters(
-#'          attributes=list("A598372E11E9910D1CBF0080EFD54D63", "A59855D811E9910D1CC50080EFD54D63"),
-#'          metrics=list("B4054F5411E9910D672E0080EFC5AE5B"),
-#'          attr_elements=list("A598372E11E9910D1CBF0080EFD54D63:Los Angeles", 
-#'          "A598372E11E9910D1CBF0080EFD54D63:Seattle"))
-#' df <- my_cube$to_dataframe()
+#' # Specify attributes and metrics (columns) to be fetched.
+#' my_cube$apply_filters(attributes = my_report$attributes[1:2],
+#'                          metrics = my_report$metrics[1:2])
+#'
+#' # See the selection of attributes, metrics and attribute elements.
+#' my_cube$selected_attributes
+#' my_cube$selected_metrics
+#' my_cube$selected_attr_elements
+#'
+#' # Clear filtering to load a full dataset.
+#' my_cube$clear_filters()
+#'
+#' # Fetch data from the Intelligence Server.
+#' my_cube$to_dataframe()
+#'
+#' # See the dataframe.
+#' my_cube$dataframe
 #' }
 #' @docType class
 #' @importFrom R6 R6Class
@@ -35,6 +48,7 @@ Cube <- R6Class("Cube",
     #instance variables
     connection = NULL,
     cube_id = NULL,
+    offset = 0,
     name = NULL,
     owner_id = NULL,
     path = NULL,
@@ -47,6 +61,10 @@ Cube <- R6Class("Cube",
     selected_attributes = NULL,
     selected_metrics = NULL,
     selected_attr_elements = NULL,
+    dataframe = NULL,
+    dataframe_list = NULL,
+    table_definition = NULL,
+    filters = NULL,
 
     initialize = function(connection, cube_id) {
       # Initialize cube contructor.
@@ -56,51 +74,27 @@ Cube <- R6Class("Cube",
 
       private$load_info()
       private$load_definition()
+
+      self$filters <- Filter$new(attributes = self$attributes,
+                                metrics = self$metrics,
+                                attr_elements = NULL)
     },
 
-    get_attr_elements = function(verbose = TRUE) {
-      # Load elements of all attributes to be accesibile by Cube$attr_elements.
-
-      if(is.null(self$attr_elements)){
-        private$load_attr_elements()
-      }
-
-      if(verbose) self$attr_elements
-    },
-
-    apply_filters = function(attributes = NULL, metrics = NULL, attr_elements = NULL, body = NULL) {
-      # Apply filters on the cube data so only the chosen attributes, metrics, and attribute elements are retrieved from the Intelligence Server.
-      if(!is.null(body)){
-        private$filters <- body
-      } else {
-        self$selected_attributes <- attributes
-        self$selected_metrics <- metrics
-        self$selected_attr_elements <- attr_elements
-        private$filters <- full_body(attributes, metrics, attr_elements)
-      }
-    },
-
-    clear_filters = function() {
-      # Clear previously set filters, allowing all attributes, metrics, and attribute elements to be retrieved from the Intelligence Server.
-
-      self$selected_attributes <- NULL
-      self$selected_metrics <- NULL
-      self$selected_attr_elements <- NULL
-      private$filters <- NULL
-
-    },
-
-    to_dataframe = function(offset = 0, limit = 25000, multi_df=FALSE, callback = function(x,y) {}) {
+    to_dataframe = function(limit = 25000, multi_df = FALSE, callback = function(x,y) {}) {
       # Extract contents of a cube into a R Data Frame. Previously `get_cube()`.
       
       # Get first cube instance
-      response <- cube_instance(connection=self$connection, cube_id=self$cube_id, offset=offset, limit=limit, body=private$filters)
+      response <- cube_instance(connection=self$connection,
+                                cube_id=self$cube_id,
+                                offset=self$offset,
+                                limit=limit,
+                                body=self$filters$filter_body())
 
       if(http_error(response)) {  # http != 200
-            private$response_handler(response, "Error getting cube contents.")
+            private$response_handler(response, private$err_msg_instance)
       }
 
-      # Gets the pagination totals from the response object
+      # Get the pagination totals from the response object
       response <- content(response)
       pagination <- response$result$data$paging
 
@@ -120,66 +114,123 @@ Cube <- R6Class("Cube",
 
         for(offset_ in offsets){
           # Fetch add'l rows from the cube
-          response <- cube_instance_id(connection=self$connection, cube_id=self$cube_id, instance_id=instance_id, offset=offset_, limit=limit)
+          response <- cube_instance_id(connection=self$connection,
+                                       cube_id=self$cube_id,
+                                       instance_id=instance_id,
+                                       offset=offset_,
+                                       limit=limit)
           response <- content(response)
           response_list[[length(response_list)+1]] <- parse_json(response = response)
           callback(offset_, pagination$total)
         }
         callback(pagination$total, pagination$total)
-        big_df <- do.call(rbind.data.frame, response_list)
+        self$dataframe <- do.call(rbind.data.frame, response_list)
       } else {
         callback(pagination$total, pagination$total)
-        big_df <- parse_json(response=response)
+        self$dataframe <- parse_json(response=response)
       }
-
+      
+      # convert column types
+      self$dataframe <- type.convert(self$dataframe, as.is = TRUE)
+      
       if(isTRUE(multi_df)) {
 
-        # save the multitable_definition response
-        tables_columns <- self$multitable_definition()
+        # save the multitable_definition response to the instance
+        private$multitable_definition()
 
         # split dataframe to dataframes matching tables in Cube
-        list_of_df <- list()
+        self$dataframe_list <- list()
 
-        table_names <- names(tables_columns)
+        table_names <- names(self$table_definition)
         for (name in table_names) {
-          list_of_df[[name]] <- subset(big_df, select = unlist(tables_columns[[name]]))
+          self$dataframe_list[[name]] <- subset(self$dataframe, select = unlist(self$table_definition[[name]]))
         }
 
-        return(list_of_df)
+        return(self$dataframe_list)
+      } else if(length(self$dataframe) == 0) {
+        stop(print("Data Frame looks empty."))
       } else {
-        return(big_df)
+      return(self$dataframe)
+    }
+    },
+
+    apply_filters = function(attributes = NULL, metrics = NULL, attr_elements = NULL) {
+      # Instantiate the filter object and download all attr_elements of the cube.
+      # Apply filters on the cube data so only the chosen attributes, metrics, 
+      # and attribute elements are retrieved from the Intelligence Server.
+
+      #1st check: if params is null finish
+      if(is.null(attributes) & is.null(metrics) & is.null(attr_elements)) {
+      }
+      else {
+        
+        #TODO replace with get_attr_elements()
+        #2nd check: if self.attr_elem is null
+        if(is.null(self$attr_elements) & !is.null(attr_elements)) {
+          #load attr_elem and save the result to report instance
+          private$load_attr_elements()
+        }
+        # 3rd check if filter object is created with attr_elem and if report has attr_elem downloaded
+        if(is.null(self$filters$attr_elems) & !is.null(attr_elements)) {
+          self$filters <- Filter$new(attributes=self$attributes,
+                                          metrics=self$metrics,
+                                          attr_elements=self$attr_elements)
+        }
+
+        # Previous functionality of apply_filters: set filters as specified in apply filters
+        if(!is.null(attributes)){
+          if(length(attributes)==0){
+            self$filters$attr_selected <- list()
+          }else{
+            self$filters$select(attributes)
+          }
+        } # Else do nothing
+
+        if(!is.null(metrics)){
+          if(length(metrics)==0){
+            self$filters$metr_selected <- list()
+          }else{
+            self$filters$select(metrics)
+          }
+        } # Else do nothing
+
+        if(length(attr_elements)>0) {
+          self$filters$select(attr_elements)
+        }
+
+        # Assign filter values in the report instance
+        self$selected_attributes <- self$filters$attr_selected
+        self$selected_metrics <- self$filters$metr_selected
+        self$selected_attr_elements <- self$filters$attr_elem_selected
       }
     },
 
-    multitable_definition = function() {
-      # Return all tables names and collumns as a list of list
+    clear_filters = function() {
+      # Clear previously set filters, allowing all attributes, metrics, and attribute elements to be retrieved from the Intelligence Server.
+      self$filters$clear()
 
-      response <- dataset_definition(connection=self$connection, dataset_id=self$cube_id)
+      self$selected_attributes <- self$filters$attr_selected
+      self$selected_metrics <- self$filters$metr_selected
+      self$selected_attr_elements <- self$filters$attr_elem_selected
+    },
 
-      if(http_error(response)) {  # http != 200
-        private$response_handler(response, "Error getting cube information.")
+    get_attr_elements = function(verbose = TRUE) {
+      # Load elements of all attributes to be accesibile by Cube$attr_elements.
+
+      if (is.null(self$attr_elements)) {
+        private$load_attr_elements()
       }
-      response <- content(response, as = "parsed", type = "application/json")
-
-      # Create the list of list from response
-      table_columns <- list()
-      for (table in response$result$definition$availableObjects$tables) {
-        column_list <- list()
-        for (column in response$result$definition$availableObjects$columns) {
-          if (table$name == column$tableName) {
-            column_list <- append(column_list, column$columnName)
-          }
-        }
-        # Add another table name and column list to the table_columns list
-        table_columns[[table$name]] = column_list
-      }
-      return(table_columns)
+      if (verbose) self$attr_elements
     }
   ),
 
+
   private = list(
 
-    filters = NULL,
+    err_msg_information = "Error getting cube information. Check cube ID.",
+    err_msg_definition = "Error getting cube definition. Check cube ID.",
+    err_msg_instance = "Error getting cube contents.",
+    err_msg_elements = "Error loading attribute elements.",
 
     load_info = function() {
       # Get metadata for specific cubes. Implements GET /cubes to retrieve basic metadata.
@@ -187,7 +238,7 @@ Cube <- R6Class("Cube",
       response <- cube_info(connection=self$connection, cube_id=self$cube_id)
 
       if(http_error(response)) {  # http != 200
-        private$response_handler(response, "Error getting cube information.")
+        private$response_handler(response, private$err_msg_information)
       }
 
       info <- content(response)$cubesInfos[[1]]
@@ -207,7 +258,7 @@ Cube <- R6Class("Cube",
       response <- cube(connection = self$connection, cube_id = self$cube_id)
 
       if(http_error(response)) {  # http != 200
-        private$response_handler(response, "Error loading Cube$attributes and Cube$metrics.")
+        private$response_handler(response, private$err_msg_definition)
       }
 
       objects <- content(response)$result$definition$availableObjects
@@ -221,23 +272,83 @@ Cube <- R6Class("Cube",
 
     load_attr_elements = function() {
       # Get the elements of cube attributes. Implements GET /cubes/<cube_id>/attributes/<attribute_id>/elements
+      is_empty <- function(x){
+      # Helper function to handle empty 'formValues'.
+        if(is.na(x) || is.null(x) || nchar(x)==0) return(TRUE)
+        else return(FALSE)
+      }
 
-      get_single_attr_elements = function(conn, cube_id, attr_id) {
-        #helper function extracting attr elements from HTTP response.
-        response <- cube_single_attr_elements(conn, cube_id, attr_id)
+      get_single_attr_elements = function(conn, cube_id, attr_id, offset=self$offset, limit=160000) {
+        # Helper function extracting attr elements from HTTP response.
+        
+        # Fetch first chunk to determine what is the total number of elements.
+        response <- cube_elements(conn, cube_id, attr_id, offset, limit)
 
         if(http_error(response)) {  # http != 200
-          private$response_handler(response, "Error loading Cube$attr_elements.")
+          private$response_handler(response, private$err_msg_elements)
         }
-
+        
         elements <- lapply(content(response), function(elem) unlist(elem$id))
-        names(elements) <- lapply(content(response), function(elem) unlist(elem$formValues))
+        names(elements) <- lapply(content(response), function(elem){
+                              if(is_empty(unlist(elem$formValues))) unlist(elem$id)
+                              else unlist(elem$formValues)
+                              })
+                                  
+        total <- as.numeric(response$headers$'x-mstr-total-count')
+      
+        # Fetch the rest of elements if their number exceeds the limit.
+        if(total > limit){
+          offsets = seq(from=limit, to=total, by=limit)
 
+          for(offset_ in offsets){
+            response <- cube_elements(conn, cube_id, attr_id, offset_, limit)
+
+            if(http_error(response)) {  # http != 200
+              private$response_handler(response, private$err_msg_elements)
+            }
+
+            chunk <- lapply(content(response), function(elem) unlist(elem$id))
+            names(chunk) <- lapply(content(response), function(elem){
+                              if(is_empty(unlist(elem$formValues))) unlist(elem$id)
+                              else unlist(elem$formValues)
+                            })
+            elements <- c(elements, chunk)
+          }
+        }
+                                
         elements
       }
 
-      self$attr_elements <- lapply(self$attributes, function(attr_id) get_single_attr_elements(self$connection, self$cube_id, attr_id))
+      # Fetch elements for all attributes
+      self$attr_elements <- lapply(self$attributes, function(attr_id){
+        list("id" = attr_id,
+             "elements" = get_single_attr_elements(self$connection, self$cube_id, attr_id))
+      })
 
+    },
+
+    multitable_definition = function() {
+      # Return all tables names and collumns as a list of list
+
+      response <- dataset_definition(connection=self$connection, dataset_id=self$cube_id)
+
+      if(http_error(response)) {  # http != 200
+        private$response_handler(response, "Error getting cube information.")
+      }
+      response <- content(response, as = "parsed", type = "application/json")
+
+      # Create the list of list from response
+      self$table_definition <- list()
+      for (table in response$result$definition$availableObjects$tables) {
+        column_list <- list()
+        for (column in response$result$definition$availableObjects$columns) {
+          if (table$name == column$tableName) {
+            column_list <- append(column_list, column$columnName)
+          }
+        }
+        # Add another table name and column list to the table_definition list
+        self$table_definition[[table$name]] <- column_list
+      }
     },
 
     response_handler = function(response, msg) {