+++ inst/www/precache-manifest.115eaefa18858d2346d14ea18996149d.js
@@ -1,0 +1,26 @@
+self.__precacheManifest = (self.__precacheManifest || []).concat([
+  {
+    "revision": "94ae4be569b9fc2c67b336e2a811125d",
+    "url": "./index.html"
+  },
+  {
+    "revision": "7e876c7d7985eaee124e",
+    "url": "./static/css/2.520b5698.chunk.css"
+  },
+  {
+    "revision": "e3b3ff977af7da4de3ca",
+    "url": "./static/css/main.32805f94.chunk.css"
+  },
+  {
+    "revision": "7e876c7d7985eaee124e",
+    "url": "./static/js/2.e1140f2f.chunk.js"
+  },
+  {
+    "revision": "e3b3ff977af7da4de3ca",
+    "url": "./static/js/main.8f1f35b9.chunk.js"
+  },
+  {
+    "revision": "5eb9fe67b5985a0ab537",
+    "url": "./static/js/runtime-main.6c141fb0.js"
+  }
+]);